Refactor: Move Macro to repo package and update references
Reorganized the `Macro` implementation by relocating it from the `command` package to the `repo` package. Updated all associated imports and references to align with the new structure. This improves package cohesion and better separates concerns.
