+++ pkg/repo/macro.go
@@ -1,4 +1,4 @@
-package command
+package repo
 
 import (
 	"bytes"
@@ -9,6 +9,7 @@ import (
 	"text/template"
 
 	"github.com/ksysoev/wsget/pkg/core"
+	"github.com/ksysoev/wsget/pkg/core/command"
 	"gopkg.in/yaml.v3"
 )
 
@@ -40,11 +41,11 @@ func NewMacro(domains []string) *Macro {
 // Otherwise, it creates a new Sequence with the commands and adds it to the macro.
 func (m *Macro) AddCommands(name string, rawCommands []string) error {
 	if _, ok := m.macro[name]; ok {
-		return &ErrDuplicateMacro{name}
+		return &command.ErrDuplicateMacro{name}
 	}
 
 	if len(rawCommands) == 0 {
-		return ErrEmptyMacro{name}
+		return command.ErrEmptyMacro{name}
 	}
 
 	templs, err := NewMacroTemplates(rawCommands)
@@ -63,7 +64,7 @@ func (m *Macro) AddCommands(name string, rawCommands []string) error {
 func (m *Macro) merge(macro *Macro) error {
 	for name, cmd := range macro.macro {
 		if _, ok := m.macro[name]; ok {
-			return &ErrDuplicateMacro{name}
+			return &command.ErrDuplicateMacro{name}
 		}
 
 		m.macro[name] = cmd
@@ -79,7 +80,7 @@ func (m *Macro) Get(name, argString string) (core.Executer, error) {
 		return cmd.GetExecuter(args)
 	}
 
-	return nil, &ErrUnknownCommand{name}
+	return nil, &command.ErrUnknownCommand{name}
 }
 
 func (m *Macro) GetNames() []string {
@@ -106,7 +107,7 @@ func LoadFromFile(path string) (*Macro, error) {
 	}
 
 	if cfg.Version != "1" {
-		return nil, &ErrUnsupportedVersion{cfg.Version}
+		return nil, &command.ErrUnsupportedVersion{cfg.Version}
 	}
 
 	macroCfg := NewMacro(cfg.Domains)
@@ -201,7 +202,7 @@ func (t *MacroTemplates) GetExecuter(args []string) (core.Executer, error) {
 			return nil, err
 		}
 
-		cmd, err := NewFactory(nil).Create(output.String())
+		cmd, err := command.NewFactory(nil).Create(output.String())
 		if err != nil {
 			return nil, err
 		}
@@ -213,5 +214,5 @@ func (t *MacroTemplates) GetExecuter(args []string) (core.Executer, error) {
 		return cmds[0], nil
 	}
 
-	return NewSequence(cmds), nil
+	return command.NewSequence(cmds), nil
 }
